+++ src/utils/calculate-overlay.tsx
@@ -1,4 +1,8 @@
-const calculateOverlay = (foreGroundRgba, backgroundRgb) => {
+const channelResult = (fg: number, bg: number, alpha:number): number => {
+  return Math.floor(alpha * fg + (1 - alpha) * bg);
+};
+
+const calculateOverlay = (foreGroundRgba: Array<number>, backgroundRgb: Array<number>): Array<number> => {
   const fgR = foreGroundRgba[0];
   const fgG = foreGroundRgba[1];
   const fgB = foreGroundRgba[2];
@@ -7,10 +11,6 @@ const calculateOverlay = (foreGroundRgba, backgroundRgb) => {
   const bgR = backgroundRgb[0];
   const bgG = backgroundRgb[1];
   const bgB = backgroundRgb[2];
-
-  const channelResult = (fg, bg, alpha) => {
-    return Math.floor(alpha * fg + (1 - alpha) * bg);
-  };
 
   const r = channelResult(fgR, bgR, fgA);
   const g = channelResult(fgG, bgG, fgA);
